Collect and separate classes to a size of module(package)
