+++ src/main/java/com/minkyu/hellospring/ObjectFactory.java
@@ -1,6 +1,7 @@
 package com.minkyu.hellospring;
 
-import com.minkyu.hellospring.exrate.ExRateProvider;
+import com.minkyu.hellospring.exrate.CachedExRateProvider;
+import com.minkyu.hellospring.payment.ExRateProvider;
 import com.minkyu.hellospring.exrate.WebApiExRatePaymentService;
 import com.minkyu.hellospring.payment.PaymentService;
 import org.springframework.context.annotation.Bean;
