+++ src/main/java/com/minkyu/hellospring/exrate/ExRateData.java
@@ -1,4 +1,4 @@
-package com.minkyu.hellospring;
+package com.minkyu.hellospring.exrate;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
